cha/requirements.txt
aiohttp
pillow
pywebio
brotlipy
pyinstaller